requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pgmpy==0.1.24
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
bcrypt>=4.1.0
google-generativeai>=0.5.0
